+++ README.md
@@ -39,7 +39,7 @@ Latest API Docs is <a target="_blank" href="https://wso2-extensions.github.io/si
 
 ```
 <dependency>
-  <groupId>org.wso2.extension.siddhi.io.cdc</groupId>
+  <groupId>io.siddhi.extension.io.cdc</groupId>
   <artifactId>siddhi-io-cdc</artifactId>
   <version>x.x.x</version>
  </dependency>
