+++ component/src/test/java/io/siddhi/extension/io/cdc/source/TestCaseOfCDCListeningMode.java
@@ -1,0 +1,334 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.siddhi.extension.io.cdc.source;
+
+import io.siddhi.core.SiddhiAppRuntime;
+import io.siddhi.core.SiddhiManager;
+import io.siddhi.core.event.Event;
+import io.siddhi.core.query.output.callback.QueryCallback;
+import io.siddhi.core.stream.input.InputHandler;
+import io.siddhi.core.stream.output.StreamCallback;
+import io.siddhi.core.util.SiddhiTestHelper;
+import org.apache.log4j.Logger;
+import org.testng.Assert;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+
+public class TestCaseOfCDCListeningMode {
+    private static final Logger log = Logger.getLogger(TestCaseOfCDCListeningMode.class);
+    private Event currentEvent;
+    private AtomicInteger eventCount = new AtomicInteger(0);
+    private AtomicBoolean eventArrived = new AtomicBoolean(false);
+    private int waitTime = 50;
+    private int timeout = 10000;
+    private String username;
+    private String password;
+    private String mysqlJdbcDriverName;
+    private String databaseURL;
+    private String tableName = "login";
+
+    @BeforeClass
+    public void initializeConnectionParams() {
+        String port = System.getenv("PORT");
+        String host = System.getenv("DOCKER_HOST_IP");
+        databaseURL = "jdbc:mysql://" + host + ":" + port + "/SimpleDB?useSSL=false";
+        username = System.getenv("DATABASE_USER");
+        password = System.getenv("DATABASE_PASSWORD");
+        mysqlJdbcDriverName = "com.mysql.jdbc.Driver";
+    }
+
+    @BeforeMethod
+    public void init() {
+        eventCount.set(0);
+        eventArrived.set(false);
+        currentEvent = new Event();
+    }
+
+    /**
+     * Test case to Capture Insert operations from a MySQL table.
+     */
+    @Test
+    public void testInsertCDC() throws InterruptedException {
+        log.info("------------------------------------------------------------------------------------------------");
+        log.info("CDC TestCase: Capturing Insert change data from MySQL.");
+        log.info("------------------------------------------------------------------------------------------------");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cdcinStreamDefinition = "@app:name('cdcTesting')" +
+                "@source(type = 'cdc'," +
+                " url = '" + databaseURL + "'," +
+                " username = '" + username + "'," +
+                " password = '" + password + "'," +
+                " table.name = '" + tableName + "', " +
+                " operation = 'insert', " +
+                " @map(type='keyvalue'))" +
+                "define stream istm (id string, name string);";
+
+        String rdbmsStoreDefinition = "define stream insertionStream (id string, name string);" +
+                "@Store(type='rdbms', jdbc.url='" + databaseURL + "'," +
+                " username='" + username + "', password='" + password + "' ," +
+                " jdbc.driver.name='" + mysqlJdbcDriverName + "')" +
+                "define table login (id string, name string);";
+
+        String rdbmsQuery = "@info(name='query2') " +
+                "from insertionStream " +
+                "insert into login;";
+
+        SiddhiAppRuntime cdcAppRuntime = siddhiManager.createSiddhiAppRuntime(cdcinStreamDefinition);
+
+        StreamCallback insertionStreamCallback = new StreamCallback() {
+            @Override
+            public void receive(Event[] events) {
+                for (Event event : events) {
+                    currentEvent = event;
+                    eventCount.getAndIncrement();
+                    log.info(eventCount + ". " + event);
+                    eventArrived.set(true);
+                }
+            }
+        };
+
+        cdcAppRuntime.addCallback("istm", insertionStreamCallback);
+        cdcAppRuntime.start();
+
+        QueryCallback rdbmsQueryCallback = new QueryCallback() {
+            @Override
+            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
+                for (Event event : inEvents) {
+                    log.info("insert done: " + event);
+                }
+            }
+        };
+
+        SiddhiAppRuntime rdbmsAppRuntime = siddhiManager.createSiddhiAppRuntime(rdbmsStoreDefinition + rdbmsQuery);
+        rdbmsAppRuntime.addCallback("query2", rdbmsQueryCallback);
+        rdbmsAppRuntime.start();
+
+        //Do an insert and wait for cdc app to capture.
+        InputHandler rdbmsInputHandler = rdbmsAppRuntime.getInputHandler("insertionStream");
+        Object[] insertingObject = new Object[]{"e001", "testEmployer"};
+        rdbmsInputHandler.send(insertingObject);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
+
+        //Assert event arrival.
+        Assert.assertTrue(eventArrived.get());
+
+        //Assert event data.
+        Assert.assertEquals(insertingObject, currentEvent.getData());
+
+        cdcAppRuntime.shutdown();
+        rdbmsAppRuntime.shutdown();
+        siddhiManager.shutdown();
+    }
+
+    /**
+     * Test case to Capture Delete operations from a MySQL table.
+     */
+    @Test
+    public void testDeleteCDC() throws InterruptedException {
+        log.info("------------------------------------------------------------------------------------------------");
+        log.info("CDC TestCase: Capturing Delete change data from MySQL.");
+        log.info("------------------------------------------------------------------------------------------------");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cdcinStreamDefinition = "@app:name('cdcTesting')" +
+                "@source(type = 'cdc'," +
+                " url = '" + databaseURL + "'," +
+                " username = '" + username + "'," +
+                " password = '" + password + "'," +
+                " table.name = '" + tableName + "', " +
+                " operation = 'delete', " +
+                " @map(type='keyvalue'))" +
+                "define stream delstm (before_id string, before_name string);";
+
+
+        String rdbmsStoreDefinition = "define stream DeletionStream (id string, name string);" +
+                "define stream InsertStream(id string, name string);" +
+                "@Store(type='rdbms', jdbc.url='" + databaseURL + "'," +
+                " username='" + username + "', password='" + password + "' ," +
+                " jdbc.driver.name='" + mysqlJdbcDriverName + "')" +
+                "define table login (id string, name string);";
+
+        String insertQuery = "@info(name='query3') " +
+                "from InsertStream " +
+                "insert into login;";
+
+        String deleteQuery = "@info(name='queryDel') " +
+                "from DeletionStream " +
+                "delete login on login.id==id and login.name==name;";
+
+        SiddhiAppRuntime cdcAppRuntime = siddhiManager.createSiddhiAppRuntime(cdcinStreamDefinition);
+
+        StreamCallback deletionStreamCallback = new StreamCallback() {
+            @Override
+            public void receive(Event[] events) {
+                for (Event event : events) {
+                    currentEvent = event;
+                    eventCount.getAndIncrement();
+                    log.info(eventCount + ". " + event);
+                    eventArrived.set(true);
+                }
+            }
+        };
+
+        cdcAppRuntime.addCallback("delstm", deletionStreamCallback);
+        cdcAppRuntime.start();
+
+        QueryCallback rdbmsQuerycallback = new QueryCallback() {
+            @Override
+            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
+                for (Event event : inEvents) {
+                    eventCount.getAndIncrement();
+                    log.info("delete done: " + event);
+                }
+            }
+        };
+
+        SiddhiAppRuntime rdbmsAppRuntime = siddhiManager.createSiddhiAppRuntime(rdbmsStoreDefinition + insertQuery
+                + deleteQuery);
+        rdbmsAppRuntime.addCallback("queryDel", rdbmsQuerycallback);
+        rdbmsAppRuntime.start();
+
+        //Do an insert first.
+        InputHandler rdbmsInputHandler = rdbmsAppRuntime.getInputHandler("InsertStream");
+        Object[] insertingObject = new Object[]{"e001", "tobeDeletedName"};
+        rdbmsInputHandler.send(insertingObject);
+
+        //wait to complete deletion
+        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
+        eventCount.getAndDecrement();
+
+        //Delete inserted row
+        rdbmsInputHandler = rdbmsAppRuntime.getInputHandler("DeletionStream");
+        Object[] deletingObject = new Object[]{"e001", "tobeDeletedName"};
+        rdbmsInputHandler.send(deletingObject);
+
+        //wait to capture the delete event.
+        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
+
+        //Assert event arrival.
+        Assert.assertTrue(eventArrived.get());
+
+        //Assert event data.
+        Assert.assertEquals(deletingObject, currentEvent.getData());
+
+        cdcAppRuntime.shutdown();
+        rdbmsAppRuntime.shutdown();
+        siddhiManager.shutdown();
+    }
+
+    /**
+     * Test case to Capture Update operations from a MySQL table.
+     */
+    @Test
+    public void testUpdateCDC() throws InterruptedException {
+        log.info("------------------------------------------------------------------------------------------------");
+        log.info("CDC TestCase: Capturing Update change data from MySQL.");
+        log.info("------------------------------------------------------------------------------------------------");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cdcinStreamDefinition = "@app:name('cdcTesting')" +
+                "@source(type = 'cdc'," +
+                " url = '" + databaseURL + "'," +
+                " username = '" + username + "'," +
+                " password = '" + password + "'," +
+                " table.name = '" + tableName + "', " +
+                " operation = 'update', " +
+                " @map(type='keyvalue'))" +
+                "define stream updatestm (before_id string, id string, before_name string, name string);";
+
+        String rdbmsStoreDefinition = "define stream UpdateStream(id string, name string);" +
+                "define stream InsertStream(id string, name string);" +
+                "@Store(type='rdbms', jdbc.url='" + databaseURL + "'," +
+                " username='" + username + "', password='" + password + "' ," +
+                " jdbc.driver.name='" + mysqlJdbcDriverName + "')" +
+                "define table login (id string, name string);";
+
+        String insertQuery = "@info(name='query3') " +
+                "from InsertStream " +
+                "insert into login;";
+
+        String updateQuery = "@info(name='queryUpdate') " +
+                "from UpdateStream " +
+                "update login on login.id==id;";
+
+        SiddhiAppRuntime cdcAppRuntime = siddhiManager.createSiddhiAppRuntime(cdcinStreamDefinition);
+
+        StreamCallback updatingStreamCallback = new StreamCallback() {
+            @Override
+            public void receive(Event[] events) {
+                for (Event event : events) {
+                    currentEvent = event;
+                    eventCount.getAndIncrement();
+                    log.info(eventCount + ". " + event);
+                    eventArrived.set(true);
+                }
+            }
+        };
+
+        cdcAppRuntime.addCallback("updatestm", updatingStreamCallback);
+        cdcAppRuntime.start();
+
+        QueryCallback queryCallback2 = new QueryCallback() {
+            @Override
+            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
+                for (Event event : inEvents) {
+                    log.info("update done: " + event);
+                }
+            }
+        };
+        SiddhiAppRuntime rdbmsAppRuntime = siddhiManager.createSiddhiAppRuntime(rdbmsStoreDefinition + insertQuery
+                + updateQuery);
+        rdbmsAppRuntime.addCallback("queryUpdate", queryCallback2);
+        rdbmsAppRuntime.start();
+
+        //Do an insert first.
+        InputHandler rdbmsInputHandler = rdbmsAppRuntime.getInputHandler("InsertStream");
+        Object[] insertingObject = new Object[]{"e001", "empName"};
+        rdbmsInputHandler.send(insertingObject);
+
+        Thread.sleep(100);
+
+        //Update inserted row.
+        rdbmsInputHandler = rdbmsAppRuntime.getInputHandler("UpdateStream");
+        Object[] updatingObject = new Object[]{"e001", "newName"};
+        rdbmsInputHandler.send(updatingObject);
+
+        //wait to capture the update event.
+        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
+
+        //Assert event arrival.
+        Assert.assertTrue(eventArrived.get());
+
+        //Assert event data.
+        Object[] expectedEventObject = new Object[]{"e001", "e001", "empName", "newName"};
+        Assert.assertEquals(expectedEventObject, currentEvent.getData());
+
+        cdcAppRuntime.shutdown();
+        rdbmsAppRuntime.shutdown();
+        siddhiManager.shutdown();
+    }
+}
